remove order field inserter
