+++ connector/src/main/scala/com/celonis/kafka/connect/transform/fields/EmbeddedKafkaMetadata.scala
@@ -1,0 +1,5 @@
+package com.celonis.kafka.connect.transform.fields
+
+final case class EmbeddedKafkaMetadata(partition: Int, offset: Long, timestamp: Long) {
+  def partitionOffset = s"${partition}_${offset}"
+}
